+++ src/modals/ChangePoolState/index.tsx
@@ -4,7 +4,7 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faLock, faUnlock } from '@fortawesome/free-solid-svg-icons';
 import { useModal } from 'contexts/Modal';
-import { PoolState } from 'contexts/Pools';
+import { PoolState } from 'types/pools';
 import { faTimesCircle } from '@fortawesome/free-regular-svg-icons';
 import { IconProp } from '@fortawesome/fontawesome-svg-core';
 import { HeadingWrapper, PaddingWrapper } from '../Wrappers';
